+++ src/torchpme/__init__.py
@@ -1,6 +1,6 @@
 import contextlib
 
-from . import calculators, lib, potentials, utils  # noqa
+from . import calculators, lib, potentials, prefactors, tuning  # noqa
 from ._version import __version__, __version_tuple__  # noqa
 from .calculators import Calculator, EwaldCalculator, P3MCalculator, PMECalculator
 from .potentials import (